requirements.txt
python-telegram-bot[job-queue]>=22.0,<23.0
aiosqlite>=0.17.0,<0.21.0
pytz>=2023.3
python-dotenv>=1.0.0,<2.0.0